requirements.txt
openai>=1.0.0
click==8.1.3
rich==13.0.0
pytest==8.3.4
python-dotenv
google-generativeai

